+++ config.js
@@ -1,0 +1,5 @@
+module.exports = {
+  PAYMENT_LOG_CHANNEL: "payment-logs",
+  PAYMENT_VERIF_CHANNEL: "payment-verif",
+  STAFF_ROLE_NAME: "pembantu raja"
+};