+++ store/payments.js
@@ -1,0 +1,2 @@
+const payments = new Map();
+module.exports = payments;